Enable packaged installation using setup.py
This also removes the dependency on config.py from googleplay.py. The
reason for that is because GooglePlayAPI constructor already accepts
values for the fields being obtained via config.py, and only the driver
wrapper scripts pass in those values from config.py.

By removing config.py dependency, it allows the package to be installed
via easy_install, which will only install the googleplay and
googleplay_pb2 modules to the system, avoiding the need to introduce any
of the other scripts.

Change-Id: Ice195c62369000ecd2778fd5cbf6714601ae338f